changed data type for pgsql in heroku groupby error fixing

+++ database/migrations/2020_07_19_154942_create_farmer_sales_view.php
@@ -20,10 +20,10 @@ class CreateFarmerSalesView extends Migration
             posts.user_id AS user_id,
             posts.crop_name AS crop_name,
             posts.created_at AS created_at,
-            SUM(posts.fixed_quantity::decimal) AS totalFixedQty,
-            SUM(posts.crop_quantity::decimal) AS totalAvailableQty,
+            SUM(posts.fixed_quantity) AS totalFixedQty,
+            SUM(posts.crop_quantity) AS totalAvailableQty,
             posts.earnings AS earnings,
-            SUM(posts.kilogram_sold::decimal) AS totalKgSold,
+            SUM(posts.kilogram_sold) AS totalKgSold,
             AVG(ROUND(posts.percentage_sold_before_harvest,
                     1)) AS totalPercentage
         FROM
